update para float no models e form e icons do bootstrap

+++ produtos/forms.py
@@ -2,7 +2,7 @@ from django import forms
 
 
 class ProdutoForm(forms.Form):
-    qtd = forms.IntegerField(label='', required=True, widget=forms.TextInput(
+    qtd = forms.FloatField(label='', required=True, widget=forms.TextInput(
         attrs={'type': 'number', 'class': 'border-1 form-control text-center qtd'}))
